backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
motor==3.6.0
pymongo==4.9.1
opencv-python==4.10.0.84
numpy==1.26.4
face_recognition==1.3.0
Pillow==10.4.0
